refactor(model):Presentation Mappers 구현
