+++ app/src/main/java/woowacourse/movie/presentation/model/TicketModel.kt
@@ -2,7 +2,6 @@ package woowacourse.movie.presentation.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import woowacourse.movie.domain.model.Ticket
 import java.time.LocalDateTime
 
 @Parcelize
@@ -10,17 +9,6 @@ data class TicketModel(
     val movieId: Long,
     val bookedDateTime: LocalDateTime,
     val count: Int,
-) : Parcelable {
+    val paymentMoney: Int,
 
-    fun toTicket() = Ticket(
-        movieId = movieId,
-        bookedDateTime = bookedDateTime,
-        count = count
-    )
-}
-
-fun Ticket.toPresentation() = TicketModel(
-    movieId = movieId,
-    bookedDateTime = bookedDateTime,
-    count = count
-)
+) : Parcelable
